More minor bug fixes: jgrapht-core/src/main/java/overview.html L4 Subsituted <b> by <strong> jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctDirectedGraph.java L125,134 long cast jgrapht-core/src/main/java/org/jgrapht/alg/tour/PalmerHamiltonianCycle.java L131,133 commented assignments from self to self jgrapht-core/src/main/java/org/jgrapht/generate/netgen/NetworkGeneratorConfig.java L292 long cast jgrapht-core/src/main/java/org/jgrapht/traverse/ClosestFirstIterator.java L179 deleted hasNext() line that wasn't being used for anything jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctIntDirectedGraph.java L121,130 long cast jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctIntUndirectedDirectedGraph.java L121,130 long cast jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctUndirectedGraph.java L119,121 long cast jgrapht-core/src/main/java/org/jgrapht/alg/shortestpath/IntVertexDijkstraShortestPath.java L273

+++ jgrapht-core/src/main/java/org/jgrapht/generate/netgen/NetworkGeneratorConfig.java
@@ -289,7 +289,7 @@ public class NetworkGeneratorConfig {
      * @return minimum number of nodes this network can contain.
      */
     public long getMinimumArcNum() {
-        return getTransshipNodeNum() + Math.max(getSourceNum(), getSinkNum());
+        return (long) (getTransshipNodeNum() + Math.max(getSourceNum(), getSinkNum()));
     }
 
     /**
